backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
bcrypt==4.1.3
boto3==1.34.144
python-dotenv==1.0.1
reportlab==4.2.2
gunicorn==22.0.0
google-cloud-billing==1.13.3
google-cloud-resource-manager==1.12.3
azure-mgmt-costmanagement==4.0.1
azure-mgmt-resource==23.0.0
azure-mgmt-compute==30.6.0
azure-mgmt-storage==21.1.0
azure-mgmt-monitor==6.0.2
azure-identity==1.16.1
cryptography==42.0.8
flask-mail==0.10.0
itsdangerous==2.2.0
flask-limiter==3.5.0
requests==2.31.0
google-cloud-compute==1.14.1
razorpay==1.4.2
